make dashhoard file list item a link
